Bugfix: Repair close on log generated when save log on launch.

+++ NWNLogRotator.xaml.cs
@@ -243,7 +243,8 @@ namespace NWNLogRotator
                     }
                     else
                     {
-                        Process.GetCurrentProcess().Kill();
+                        if(_settings.CloseOnLogGenerated == true )
+                            Process.GetCurrentProcess().Kill();
                     }
                         
                 }
